Move installer stuff to separate module
