+++ src/lib.rs
@@ -3,6 +3,7 @@ use url::Url;
 
 pub mod chromedriver;
 pub mod geckodriver;
+pub mod installer;
 
 pub trait DriverFetcher {
     const BASE_URL: &'static str;
